requirements.txt
again==1.2.21
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
blinker==1.7.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.2
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==8.3.0
orjson==3.10.5
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.4
pydantic-extra-types==2.8.2
pydantic-settings==2.3.3
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.1
